+++ tests/E2E Tests/TokenAcquirerTests/TokenAcquirer.cs
@@ -308,6 +308,7 @@ namespace TokenAcquirerTests
         //[Fact]
         public async Task AcquireTokenWithManagedIdentity_UserAssigned()
         {
+            // Arrange
             const string scope = "https://vault.azure.net/.default";
             const string baseUrl = "https://vault.azure.net";
             const string clientId = "9c5896db-a74a-4b1a-a259-74c5080a3a6a";
@@ -315,12 +316,12 @@ namespace TokenAcquirerTests
             _ = tokenAcquirerFactory.Services;
             IServiceProvider serviceProvider = tokenAcquirerFactory.Build();
 
-            // Get the authorization header provider and add the options to tell it to use Managed Identity
+            // Act: Get the authorization header provider and add the options to tell it to use Managed Identity
             IAuthorizationHeaderProvider? api = serviceProvider.GetRequiredService<IAuthorizationHeaderProvider>();
             Assert.NotNull(api);
             string result = await api.CreateAuthorizationHeaderForAppAsync(scope, GetAuthHeaderOptions_ManagedId(baseUrl, clientId));
 
-            // Make sure we got a token
+            // Assert: Make sure we got a token
             Assert.False(string.IsNullOrEmpty(result));
         }
